requirements.txt
# already in streamlit env
streamlit~=1.41.1

requests==2.32.3
urllib3==2.2.2
st-pages==1.0.1
pandas~=2.2.3
plotly~=5.24.1
alembic~=1.15.2
psycopg2-binary~=2.9.10
sqlalchemy~=2.0.40
toml~=0.10.2
pycryptodome~=3.22.0
streamlit_local_storage~=0.0.25
streamlit-extras~=0.6.0
